meta: add changelog entry in docs/changelog per CI requirements
