+++ server/src/main/java/org/elasticsearch/index/mapper/vectors/VectorSimdOptimizer.java
@@ -1,44 +1,37 @@
 package org.elasticsearch.index.mapper.vectors;
 
-import jdk.incubator.vector.FloatVector;
-import jdk.incubator.vector.VectorSpecies;
-import jdk.incubator.vector.VectorOperators;  
+import java.lang.reflect.Method;
 
+public class VectorSimdOptimizer {
+    private static final Method SIMD_METHOD;
 
-public final class VectorSimdOptimizer {            
-
-    
-    private static final VectorSpecies<Float> SPECIES = FloatVector.SPECIES_PREFERRED;
-
-    private VectorSimdOptimizer() {
-        
+    static {
+        Method method = null;
+        try {
+            Class<?> clazz = Class.forName("org.elasticsearch.index.mapper.vectors.VectorSimdOptimizerJava21");
+            method = clazz.getMethod("dotProductSIMD", float[].class, float[].class);
+        } catch (Throwable t) {
+            method = null;
+        }
+        SIMD_METHOD = method;
     }
 
-    
-    public static float dotProductSIMD(float[] a, float[] b) {
-        int length = a.length;
-        int loopBound = SPECIES.loopBound(length);
-        
-        
-        FloatVector acc = FloatVector.zero(SPECIES);
-
-        int i = 0;
-        
-        for (; i < loopBound; i += SPECIES.length()) {
-            FloatVector va = FloatVector.fromArray(SPECIES, a, i);
-            FloatVector vb = FloatVector.fromArray(SPECIES, b, i);
-            acc = va.fma(vb, acc); 
+    public static float dotProduct(float[] a, float[] b) {
+        if (SIMD_METHOD != null) {
+            try {
+                return (float) SIMD_METHOD.invoke(null, (Object) a, (Object) b);
+            } catch (Throwable t) {
+                // Fallback silencioso se falhar a invocação por segurança
+            }
         }
+        return dotProductScalar(a, b);
+    }
 
-        
-       
-        float totalSum = acc.reduceLanes(VectorOperators.ADD);  
-
-        
-        for (; i < length; i++) {
-            totalSum += a[i] * b[i];
+    private static float dotProductScalar(float[] a, float[] b) {
+        float res = 0f;
+        for (int i = 0; i < a.length; i++) {
+            res += a[i] * b[i];
         }
-
-        return totalSum;
+        return res;
     }
 }
